Struct humanize::matchers::Match

pub struct Match {
    pub value: HumanValue,
    pub weight: i32,
}

A possible match for a value within some text.

A Match result is obtained by calling parse on some input text. They are created by Matchers.

Fields

value

The value determined for this match.

weight

Strength of the match.

This is useful when there is more than one possible match.

TODO: Should be this be a percentage? What is the range? Should it be an enum with values like 'Likely', 'Unlikely', and 'Certain'?