Module humanize::parse

Human-friendly Parsing

This module provides support for finding human-friendly values within some source text. This is typically used for parsing input text for prompts that need common data types such as instants in time, durations, booleans, etc.

Upcoming Breaking Changes

The system of using ValueType and HumanValue as seen in the current implementation of this library will be going away in favor of something else (and depending on another crate) in the near future so that we can interoperate with other things that need a richer type system.

Match Input

First, you'll want to construct a parser:

use humanize::parse::Parser;

let parser = Parser::new();

Then, you can use that parser to examine some input. You'll want to specify what sort of value you are looking for. You may also limit the matchers run to a specific language. (Here, we don't limit the languages, so we pass Default::default().)

let matches = parser.parse("on", ValueType::Boolean, Default::default());
assert_eq!(matches.len(), 1);
assert_eq!(matches[0].value, HumanValue::Boolean(true));

Register Matchers

Matchers can be provided to augment the built-in capabilities of this library.

We will expand upon this in the future once our own infrastructure for doing matchers well is in place.

...

Modules

english
English Humanization

Structs

Match
A possible match for a value within some text.

Matcher

Parser

SourceLocation
A location within a body of text.

Enums

HumanValue

ValueType