Struct humanize::HumanizedParser

pub struct HumanizedParser<'p> {
    // some fields omitted
}

Parse text input by a human and return a primitive value.

The actual matching is done by Matcher objects which are installed by default when the parser is created. A parser without pre-installed matchers can be created with new_without_default_matchers. Additional or custom matchers can be added with register_matcher.

Methods

impl<'p> HumanizedParser<'p>

fn new() -> Self

Construct a new parser, including the default matchers.

fn new_without_default_matchers() -> Self

Construct a new parser, but without any of the default matchers.

fn parse_boolean(&self, text: &str) -> Option<bool>

Parse text, looking for a bool value.

fn parse(&self, text: &str, desired: ValueType) -> Vec<Match>

Parse text, looking for a value of the desired type.

The resulting collection of matches will be ordered by their weight of likelihood with the most likely first.

You won't typically invoke this directly, instead preferring to use type-specific methods like parse_boolean.

fn register_matcher(&mut self, matcher: Matcher<'p>) -> bool

Install a new Matcher to be used by this parser.

If the matcher's language does not match the parser's language, it will not be added.

The return type indicates whether or not the matcher was added to the parser.